+++ src/components/BackToTop/components/ArrowUpward.js
@@ -1,8 +1,18 @@
 import React, { memo } from 'react'
 import { SvgIcon } from '@material-ui/core'
+import { makeStyles } from '@material-ui/core/styles'
 
-export default memo(props => (
-  <SvgIcon {...props}>
-    <path d="M4 12l1.41 1.41L11 7.83V20h2V7.83l5.58 5.59L20 12l-8-8-8 8z" />
-  </SvgIcon>
-))
+const useStyles = makeStyles(({ palette }) => ({
+  icon: {
+    fill: palette.primary.contrastText,
+  },
+}))
+
+export default memo(({ className, ...rest }) => {
+  const classes = useStyles()
+  return (
+    <SvgIcon className={`${className} ${classes.icon}`} {...rest}>
+      <path d="M4 12l1.41 1.41L11 7.83V20h2V7.83l5.58 5.59L20 12l-8-8-8 8z" />
+    </SvgIcon>
+  )
+})