+++ src/components/CookiesBanner/CookiesBanner.js
@@ -1,6 +1,7 @@
 import React, { forwardRef } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
+import { makeStyles } from '@material-ui/core/styles'
 
 import Paragraph from '../Paragraph'
 
@@ -8,11 +9,23 @@ import CookiesButton from './components/CookiesButton'
 import CookieIcon from './components/CookieIcon'
 import Container from './components/Container'
 
+const useStyles = makeStyles(({ palette }) => ({
+  text: {
+    color: palette.primary.contrastText,
+  },
+}))
+
 const CookiesBanner = ({ children, className, forwardedRef = null, text = null, ...rest }) => {
+  const classes = useStyles()
+
   return (
     <Container className={className} ref={forwardedRef} {...rest}>
       <CookieIcon />
-      {text && <Paragraph color="secondary">{text}</Paragraph>}
+      {text && (
+        <Paragraph variant="body2" className={classes.text}>
+          {text}
+        </Paragraph>
+      )}
       {children}
     </Container>
   )