+++ src/components/CookiesConsent/components/MainTitle.js
@@ -3,11 +3,11 @@ import { Typography } from '@material-ui/core'
 
 export default styled(Typography).attrs(() => ({
   variant: 'h6',
-  color: 'secondary',
 }))`
   text-align: center;
   padding: 12px;
   ${({ theme }) => `
+    color: ${theme.palette.primary.contrastText};
     background-color: ${theme.palette.primary.main};
     border-radius: ${theme.shape.borderRadius}px;
     border-bottom-left-radius: 0px;