+++ src/components/Footer/components/Title.js
@@ -4,17 +4,18 @@ import styled from 'styled-components'
 import { Typography } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
 
-const useStyles = makeStyles({
+const useStyles = makeStyles(({ palette }) => ({
   title: {
     fontWeight: 500,
+    color: palette.primary.contrastText,
   },
-})
+}))
 
 const Separator = styled.hr`
   width: 30%;
   margin-bottom: 48px;
   ${({ theme }) => `
-    border-color: ${theme.palette.secondary.main};
+    border-color: ${theme.palette.primary.contrastText};
   `}
 `
 
@@ -23,7 +24,7 @@ const Title = ({ text = '' }) => {
 
   return (
     <>
-      <Typography align="center" className={classes.title} color="secondary" variant="h4">
+      <Typography align="center" className={classes.title} variant="h4">
         {text}
       </Typography>
       <Separator />